refactor(runtime): create runtime/ + rename daemon_protocol → wire_protocol (slice 4/8, #170)
