+++ pyrightconfig.json
@@ -13,14 +13,14 @@
     "src/voicecli/overlay_audio.py",
     "src/voicecli/nats",
     "src/voicecli/adapters/nats",
-    "src/voicecli/daemon_protocol.py",
+    "src/voicecli/runtime/wire_protocol.py",
     "src/voicecli/nats_stt_client.py",
-    "src/voicecli/daemon.py",
-    "src/voicecli/listen.py",
-    "src/voicecli/model_registry.py",
+    "src/voicecli/runtime/daemon.py",
+    "src/voicecli/runtime/listen.py",
+    "src/voicecli/runtime/model_registry.py",
     "src/voicecli/models.py",
-    "src/voicecli/stt_daemon.py",
-    "src/voicecli/transcribe.py",
+    "src/voicecli/runtime/stt_daemon.py",
+    "src/voicecli/runtime/transcribe.py",
     "src/voicecli/cli_doctor.py"
   ],
   "reportMissingImports": "error",
