+++ src/voicecli/overlay.py
@@ -46,7 +46,7 @@ from voicecli.overlay_draw import (  # noqa: E402
     hotkey_badge,
 )
 from voicecli.stt_client import SOCKET_PATH, send_status  # noqa: E402
-from voicecli.stt_daemon import LEVEL_FILE  # noqa: E402
+from voicecli.runtime.stt_daemon import LEVEL_FILE  # noqa: E402
 
 
 class WaveformOverlay:
