+++ src/voicecli/runtime/__init__.py
@@ -1,0 +1,5 @@
+"""voicecli.runtime — daemon, STT, model-registry, and wire-protocol modules."""
+
+from voicecli.runtime.transcribe import TranscriptionResult
+
+__all__ = ["TranscriptionResult"]
