+++ src/voicecli/runtime/model_registry.py
