+++ src/voicecli/runtime/transcribe.py
