+++ src/voicecli/runtime/wire_protocol.py
