+++ tools/file_exemptions.txt
@@ -4,6 +4,6 @@
 src/voicecli/cli.py https://github.com/Roxabi/voiceCLI/issues/129
 src/voicecli/cli_dictate.py https://github.com/Roxabi/voiceCLI/issues/129
 src/voicecli/api/api.py https://github.com/Roxabi/voiceCLI/issues/81
-src/voicecli/stt_daemon.py https://github.com/Roxabi/voiceCLI/issues/82
-src/voicecli/daemon_protocol.py https://github.com/Roxabi/voiceCLI/issues/83
-src/voicecli/transcribe.py https://github.com/Roxabi/voiceCLI/issues/83
+src/voicecli/runtime/stt_daemon.py https://github.com/Roxabi/voiceCLI/issues/82
+src/voicecli/runtime/wire_protocol.py https://github.com/Roxabi/voiceCLI/issues/83
+src/voicecli/runtime/transcribe.py https://github.com/Roxabi/voiceCLI/issues/83
